feat(dal,si-pkg,sdf): Isolate workspace export and import code
