+++ lib/dal/src/workspace.rs
@@ -1,16 +1,28 @@
 use chrono::{DateTime, Utc};
+use petgraph::Direction;
 use serde::{Deserialize, Serialize};
 use si_data_nats::NatsError;
 use si_data_pg::{PgError, PgRow};
+use si_events::ContentHash;
+use si_layer_cache::db::serialize;
+use si_layer_cache::LayerDbError;
+use si_pkg::{
+    WorkspaceExport, WorkspaceExportChangeSetV0, WorkspaceExportContentV0,
+    WorkspaceExportMetadataV0,
+};
+use std::collections::{HashMap, VecDeque};
+use std::sync::Arc;
 use telemetry::prelude::*;
 use thiserror::Error;
+use ulid::Ulid;
 
 use crate::change_set::{ChangeSet, ChangeSetError, ChangeSetId};
+use crate::layer_db_types::ContentTypes;
 use crate::workspace_snapshot::WorkspaceSnapshotError;
 use crate::{
     pk, standard_model, standard_model_accessor_ro, DalContext, HistoryActor, HistoryEvent,
     HistoryEventError, KeyPairError, StandardModelError, Tenancy, Timestamp, TransactionsError,
-    UserError, WorkspaceSnapshot,
+    User, UserError, UserPk, WorkspaceSnapshot,
 };
 
 const WORKSPACE_GET_BY_PK: &str = include_str!("queries/workspace/get_by_pk.sql");
@@ -28,10 +40,18 @@ pub enum WorkspaceError {
     ChangeSet(#[from] ChangeSetError),
     #[error("change set not found by id: {0}")]
     ChangeSetNotFound(ChangeSetId),
+    #[error("Trying to export from system actor. This can only be done by a user actor")]
+    ExportingFromSystemActor,
     #[error(transparent)]
     HistoryEvent(#[from] HistoryEventError),
+    #[error("Trying to import a changeset that does not have a valid base: {0}")]
+    ImportingOrphanChangeset(ChangeSetId),
+    #[error("invalid user {0}")]
+    InvalidUser(UserPk),
     #[error(transparent)]
     KeyPair(#[from] KeyPairError),
+    #[error("LayerDb error: {0}")]
+    LayerDb(#[from] LayerDbError),
     #[error(transparent)]
     Nats(#[from] NatsError),
     #[error("no user in context")]
@@ -306,6 +326,189 @@ impl Workspace {
         } else {
             Ok(None)
         }
+    }
+
+    pub async fn generate_export_data(
+        &self,
+        ctx: &DalContext,
+        workspace_version: &str,
+    ) -> WorkspaceResult<WorkspaceExport> {
+        let mut content_hashes = vec![];
+        let mut change_sets: HashMap<Ulid, Vec<WorkspaceExportChangeSetV0>> = HashMap::new();
+        let mut default_change_set_base = Ulid::nil();
+        for change_set in ChangeSet::list_open(ctx).await? {
+            let snap = WorkspaceSnapshot::find_for_change_set(ctx, change_set.id).await?;
+
+            // From root, get every value from every node, store with hash
+            let mut queue = VecDeque::from([snap.root().await?]);
+
+            while let Some(this_node_idx) = queue.pop_front() {
+                // Queue contents
+                content_hashes.extend(
+                    snap.get_node_weight(this_node_idx)
+                        .await?
+                        .content_store_hashes(),
+                );
+
+                let children = snap
+                    .edges_directed_by_index(this_node_idx, Direction::Outgoing)
+                    .await?
+                    .into_iter()
+                    .map(|(_, _, target)| target)
+                    .collect::<VecDeque<_>>();
+
+                queue.extend(children)
+            }
+
+            let base_changeset = change_set
+                .base_change_set_id
+                .map(|id| id.into_inner())
+                .unwrap_or(Ulid::nil());
+
+            if change_set.id == self.default_change_set_id() {
+                default_change_set_base = base_changeset
+            }
+
+            change_sets
+                .entry(base_changeset)
+                .or_default()
+                .push(WorkspaceExportChangeSetV0 {
+                    id: change_set.id.into_inner(),
+                    name: change_set.name.clone(),
+                    base_change_set_id: change_set.base_change_set_id.map(|id| id.into_inner()),
+                    workspace_snapshot_serialized_data: snap.serialized().await?,
+                })
+        }
+
+        let store_values_map = ctx
+            .layer_db()
+            .cas()
+            .read_many(content_hashes.as_ref())
+            .await?
+            .into_iter()
+            .map(|(hash, content)| (hash, (content, "postcard".to_string())))
+            .collect::<HashMap<_, _>>();
+
+        let content_store_values = serialize::to_vec(&store_values_map)?;
+
+        let created_by = if let HistoryActor::User(user_pk) = ctx.history_actor() {
+            let user = User::get_by_pk(ctx, *user_pk)
+                .await?
+                .ok_or(WorkspaceError::InvalidUser(*user_pk))?;
+
+            user.email().clone()
+        } else {
+            "SystemInit".to_string()
+        };
+
+        let metadata = WorkspaceExportMetadataV0 {
+            name: self.name().clone(),
+            version: workspace_version.to_string(),
+            description: "Workspace Backup".to_string(), // TODO Get this from the user
+            created_at: Default::default(),
+            created_by,
+            default_change_set: self.default_change_set_id().into_inner(),
+            default_change_set_base,
+            workspace_pk: self.pk().into_inner(),
+            workspace_name: self.name().clone(),
+        };
+
+        Ok(WorkspaceExport::new(WorkspaceExportContentV0 {
+            change_sets,
+            content_store_values,
+            metadata,
+        }))
+    }
+
+    pub async fn import(
+        &mut self,
+        ctx: &DalContext,
+        workspace_data: WorkspaceExport,
+    ) -> WorkspaceResult<()> {
+        let WorkspaceExportContentV0 {
+            change_sets,
+            content_store_values,
+            metadata,
+        } = workspace_data.into_latest();
+
+        // ABANDON PREVIOUS CHANGESETS
+        for mut change_set in ChangeSet::list_open(ctx).await? {
+            change_set.abandon(ctx).await?;
+        }
+
+        let base_changeset_for_default = {
+            let changeset_id = self.default_change_set_id();
+
+            let changeset = ChangeSet::find(ctx, changeset_id)
+                .await?
+                .ok_or(WorkspaceError::ChangeSetNotFound(changeset_id))?;
+
+            changeset.base_change_set_id
+        };
+
+        // Go from head changeset to children, creating new changesets and updating base references
+        let mut base_change_set_queue = VecDeque::from([metadata.default_change_set_base]);
+        let mut change_set_id_map = HashMap::new();
+        while let Some(base_change_set_ulid) = base_change_set_queue.pop_front() {
+            let Some(change_sets) = change_sets.get(&base_change_set_ulid) else {
+                continue;
+            };
+
+            for change_set_data in change_sets {
+                let imported_snapshot = WorkspaceSnapshot::from_bytes(
+                    &change_set_data.workspace_snapshot_serialized_data,
+                )
+                .await?;
+
+                // If base_change_set is default_change_set_base, it pointed to the builtin workspace
+                // originally, so this change set needs to be the new default for the workspace - HEAD
+                let mut is_new_default = false;
+                let actual_base_changeset: Option<ChangeSetId> =
+                    if base_change_set_ulid == metadata.default_change_set_base {
+                        is_new_default = true;
+                        base_changeset_for_default
+                    } else {
+                        Some(*change_set_id_map.get(&base_change_set_ulid).ok_or(
+                            WorkspaceError::ImportingOrphanChangeset(base_change_set_ulid.into()),
+                        )?)
+                    };
+
+                let mut new_change_set =
+                    ChangeSet::new(ctx, change_set_data.name.clone(), actual_base_changeset)
+                        .await?;
+
+                change_set_id_map.insert(change_set_data.id, new_change_set.id);
+
+                let new_snap_address = imported_snapshot
+                    .write(ctx, new_change_set.vector_clock_id())
+                    .await?;
+
+                new_change_set.update_pointer(ctx, new_snap_address).await?;
+
+                // Set new default changeset for workspace
+                if is_new_default {
+                    self.update_default_change_set_id(ctx, new_change_set.id)
+                        .await?;
+                }
+
+                base_change_set_queue.push_back(change_set_data.id)
+            }
+        }
+
+        let cas_values: HashMap<ContentHash, (Arc<ContentTypes>, String)> =
+            serialize::from_bytes(&content_store_values)?;
+
+        let layer_db = ctx.layer_db();
+
+        // TODO use the serialization format to ensure we're hashing the data correctly, if we change the format
+        for (_, (content, _serialization_format)) in cas_values {
+            layer_db
+                .cas()
+                .write(content, None, ctx.events_tenancy(), ctx.events_actor())
+                .await?;
+        }
+
+        Ok(())
     }
 
     standard_model_accessor_ro!(name, String);
